Move matrix inversion to graphics and clean up look_at matrix
Initially I made a mistake by getting a bit confused at some
refactoring.
There is a matrix that is called point_at and it's inverse
which is usually called look_at. For some reason every tutorial
seems to mix up the terminology.
At the moment, unfortunatelly I don't understand the concepts exactly,
but went with the current naming to make it a bit clearer for now.
This should be cleaned up a bit better.
